requirements.txt
Django
psycopg2-binary
pytest
Pillow
djangorestframework
pytest-django
django-environ
faker
